backend/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.40,<2.1

# Async task queue
Celery==5.3.4
redis==5.0.1
gunicorn==22.0.0
eventlet==0.36.1

# Multi-Provider AI Integration
groq==0.4.2
google-generativeai==0.3.0

huggingface-hub==0.19.4

# GitHub API
PyGithub==2.1.1

# Export generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==8.3.2
pytest-cov==5.0.0
